Move FreeVCConfig to TTS.vc.configs (like all other config classes)
